Integrate advanced logging features throughout codebase
Phase 1: Consolidation
- Merged loguru_advanced.py features into logging_utils.py
- Removed demonstration files to keep repository clean
- Updated CLAUDE.md with logging documentation

Phase 2: BaseTrainer Integration
- Added structured logging with context binding for epochs/steps
- Implemented ProgressTracker for training progress
- Added MetricsLogger for structured metrics output to JSONL
- Used log_timing for performance monitoring
- Applied catch_and_log decorator to evaluate method

Phase 3: Model Factory Enhancement
- Added @catch_and_log decorators for error handling
- Implemented log_timing for model creation and weight loading
- Added lazy_debug for expensive config/parameter logging
- Created parameter breakdown helper for detailed stats

Key improvements:
- Consistent structured logging across training pipeline
- Better performance visibility with automatic timing
- Reduced log noise with lazy evaluation and frequency logging
- Structured metrics output for analysis and MLflow integration
- Comprehensive error context for debugging
