+++ models/factory.py
@@ -19,6 +19,14 @@ import mlx.nn as nn
 import pandas as pd
 from loguru import logger
 
+# Import advanced logging features
+from utils.logging_utils import (
+    catch_and_log,
+    log_timing,
+    lazy_debug,
+    bind_context
+)
+
 # Import BERT configs and classes
 # Import Classic BERT architecture
 # Import ModernBERT architecture
@@ -121,6 +129,11 @@ class CompetitionAnalysis:
             self.text_columns = []
 
 
+@catch_and_log(
+    ValueError,
+    "Model creation failed",
+    reraise=True
+)
 def create_model(
     model_type: ModelType = "bert_with_head",
     config: dict[str, Any] | BertConfig | ModernBertConfig | None = None,
@@ -143,6 +156,19 @@ def create_model(
     Returns:
         Initialized model
     """
+    # Create logger with context
+    log = bind_context(model_type=model_type, head_type=head_type)
+    log.info(f"Creating model: {model_type}")
+    
+    # Use lazy debug for expensive config serialization
+    lazy_debug(
+        "Model configuration",
+        lambda: {
+            "config": config.__dict__ if hasattr(config, "__dict__") else config,
+            "head_config": head_config.__dict__ if hasattr(head_config, "__dict__") else head_config,
+            "kwargs": kwargs
+        }
+    )
     # Determine config type and create appropriate config
     if model_type in ["modernbert_core", "modernbert_with_head"]:
         # ModernBERT models
@@ -254,9 +280,38 @@ def create_model(
 
     # Load pretrained weights if provided
     if pretrained_path:
-        load_pretrained_weights(model, pretrained_path)
+        with log_timing("load_pretrained_weights", path=str(pretrained_path)):
+            load_pretrained_weights(model, pretrained_path)
+    
+    # Log model statistics
+    if hasattr(model, "num_parameters"):
+        param_count = model.num_parameters()
+        log.info(f"Model created with {param_count:,} parameters")
+        
+        # Lazy debug for parameter breakdown
+        lazy_debug(
+            "Parameter breakdown",
+            lambda: _get_parameter_breakdown(model)
+        )
 
     return model
+
+
+def _get_parameter_breakdown(model) -> dict[str, int]:
+    """Get parameter count breakdown by component."""
+    breakdown = {}
+    
+    if hasattr(model, "bert") and hasattr(model, "head"):
+        # BertWithHead model
+        if hasattr(model.bert, "num_parameters"):
+            breakdown["bert"] = model.bert.num_parameters()
+        if hasattr(model.head, "num_parameters"):
+            breakdown["head"] = model.head.num_parameters()
+    
+    # Add more component breakdowns as needed
+    breakdown["total"] = model.num_parameters() if hasattr(model, "num_parameters") else 0
+    
+    return breakdown
 
 
 def create_model_with_lora(
@@ -545,6 +600,11 @@ def create_multi_adapter_model(
     return base_model, manager
 
 
+@catch_and_log(
+    Exception,
+    "Failed to load model from checkpoint",
+    reraise=True
+)
 def create_model_from_checkpoint(checkpoint_path: str | Path) -> nn.Module:
     """
     Create and load a model from a checkpoint directory.
@@ -556,9 +616,13 @@ def create_model_from_checkpoint(checkpoint_path: str | Path) -> nn.Module:
         Loaded model
     """
     checkpoint_path = Path(checkpoint_path)
+    log = bind_context(checkpoint=str(checkpoint_path))
+    
+    with log_timing("load_model_from_checkpoint", checkpoint=str(checkpoint_path)):
+        log.info(f"Loading model from checkpoint: {checkpoint_path}")
 
-    # Look for metadata.json instead of config.json
-    metadata_path = checkpoint_path / "metadata.json"
+        # Look for metadata.json instead of config.json
+        metadata_path = checkpoint_path / "metadata.json"
 
     # Try to find training configuration in parent directories
     training_config_path = None
